# gen/http/openapi.yaml
swagger: "2.0"
info:
    title: Calculator Service
    description: Service for multiplying numbers, a Goa teaser
    version: ""
host: localhost:8000
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
    /multiply/{a}/{b}:
        get:
            tags:
                - calc
            summary: Integer multiplication
            description: Multiply two integers a and b and get the result in the response's body.
            externalDocs:
                description: Multiplication documentation
                url: http://there-is-no-multiplication-docs.com
            operationId: calc#multiply
            produces:
                - text/plain
            parameters:
                - name: a
                  in: path
                  description: First operand of multiplication payload
                  required: true
                  type: integer
                - name: b
                  in: path
                  description: Second operand of multiplication payload
                  required: true
                  type: integer
            responses:
                "200":
                    description: Successful integer multiplication response.
                    schema:
                        type: string
                "400":
                    description: Bad Request response.
                    schema:
                        $ref: '#/definitions/CalcMultiplyBadRequestResponseBody'
                "500":
                    description: Returned when the service fails to calculate the outcome of an operation.
                    schema:
                        type: string
                "504":
                    description: Gateway Timeout response.
            schemes:
                - http
    /openapi.json:
        get:
            tags:
                - docs
            summary: Download ./gen/http/openapi.json
            description: JSON document containg the API's OpenAPI 2.0 definition.
            operationId: docs#/openapi.json
            responses:
                "200":
                    description: File downloaded
                    schema:
                        type: file
            schemes:
                - http
    /openapi3.json:
        get:
            tags:
                - docs
            summary: Download ./gen/http/openapi3.json
            description: JSON document containg the API's OpenAPI 3.0 definition.
            operationId: docs#/openapi3.json
            responses:
                "200":
                    description: File downloaded
                    schema:
                        type: file
            schemes:
                - http
definitions:
    CalcMultiplyBadRequestResponseBody:
        title: 'Mediatype identifier: application/json; view=default'
        type: object
        properties:
            message:
                type: string
                example: kapoio minima sfalmatos
            name:
                type: string
                example: kapoio name sfalmatos
            occured_at:
                type: string
                example: "2012-12-22T00:59:14Z"
                format: date-time
        description: multiply_bad_request_response_body result type (default view)
        example:
            message: kapoio minima sfalmatos
            name: kapoio name sfalmatos
            occured_at: "2013-01-29T16:27:28Z"
        required:
            - name
            - message
            - occured_at